refactor(parser): split project-units symbol tables into a sibling module
Post-rebase project-units.ts crossed its 523-line module-size budget by 3
lines from this PR's EntityIndexLike split. Move the SI/conversion/currency
symbol tables and the measure-type lookup table (no behavior change) into
project-units-symbols.ts to bring it back under budget.
